decimal types changed to double

+++ UserService/Controllers/ViewModels/Common/Position.cs
@@ -2,9 +2,9 @@ namespace UserService.Controllers.ViewModels.Common
 {
     public class Position
     {
-        public decimal X { get; set; }
-        public decimal Y { get; set; }
-        public decimal Z { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Z { get; set; }
 
         public Domains.ValueObject.Position ToModel()
         {
